+++ CHANGELOG.md
@@ -4,6 +4,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [Unreleased]
+
+### Added
+
+- Add `MailMessage` `to` method for setting multiple recipients.
+
+### Changed
+
+- Update email footer: use domain-only URL, wrap tech info (ref, generator, IP, timezone) in `samp` tags.
+
 ## [5.10.2] - 2026-03-31
 
 ### Fixed
@@ -1435,6 +1445,7 @@ _Stable release based on [0.1.0-rc.1]._
 
 ## [0.0.0] - 2024-09-12
 
+[Unreleased]: https://https://github.com/internetguru/laravel-common/compare/staging...dev
 [5.10.2]: https://https://github.com/internetguru/laravel-common/compare/v5.10.1...v5.10.2
 [5.10.1]: https://https://github.com/internetguru/laravel-common/compare/v5.10.0...v5.10.1
 [5.10.0]: https://https://github.com/internetguru/laravel-common/compare/v5.9.1...v5.10.0